refactor: migrate marketing pages to new PageLayout system [REC-001]
